adding custom keymap for rev.4 writerDeck mode

+++ micro-journal-rev-4-esp32/src/keyboard/Keypad/Keypad.h
@@ -3,4 +3,4 @@
 #include <Arduino.h>
 
 void keypad_load_config(String filename, int *layers, int num_cols);
-uint8_t keypad_convert_wp_keys(String key);
+int keypad_convert_wp_keys(String key);
